+++ test/test_kinematics.py
@@ -425,7 +425,9 @@ class TestKinematics(EnvironmentSetup):
                         torques = body.ComputeInverseDynamics(dofaccel)
                         # how to verify these torques?
                         # compute M(q) and check if output torques form a symmetric matrix
-                        torquebase = body.ComputeInverseDynamics(zeros(body.GetDOF()))
+                        torquebase = body.ComputeInverseDynamics(None)
+                        
+                        testtorque_m, testtorque_c, testtorque_e = body.ComputeInverseDynamics(dofaccel,None,returncomponents=True)
                         
                         M = []
                         for i in range(body.GetDOF()):
@@ -441,22 +443,29 @@ class TestKinematics(EnvironmentSetup):
                             # should be constant with changes to velocity
                             assert(transdist(M,Mreal) <= 1e-14*M.shape[0]**2)
                         assert(transdist(dot(dofaccel,M)+torquebase,torques) <= 1e-13*len(torquebase))
+                        assert(transdist(dot(dofaccel,M),testtorque_m) <= 1e-13*len(torquebase))
                         
                         env.GetPhysicsEngine().SetGravity([0,0,0])
-                        torquenogravity = body.ComputeInverseDynamics(zeros(body.GetDOF()))
+                        torquenogravity = body.ComputeInverseDynamics(None)
+                        testtorque_m2, testtorque_c2, testtorque_e2 = body.ComputeInverseDynamics(dofaccel,None,returncomponents=True)
                         body.SetDOFVelocities(dofvelnew,[0,0,0],[0,0,0],checklimits=True)
-                        torquenoexternal = body.ComputeInverseDynamics(zeros(body.GetDOF())) # base link angular velocity exerts external torque
-                        body.SetDOFVelocities(dofvelnew,*link0vel,checklimits=True)
+                        torquenoexternal = body.ComputeInverseDynamics(None) # base link angular velocity exerts external torque
+                        testtorque_m3, testtorque_c3, testtorque_e3 = body.ComputeInverseDynamics(dofaccel,None,returncomponents=True)
+                        assert( all(abs(testtorque_e3)<=1e-10) )
+                        assert( transdist(testtorque_m, testtorque_m2) <= 1e-10 )
+                        assert( transdist(testtorque_m, testtorque_m3) <= 1e-10 )
+                        assert( transdist(testtorque_c, torquenoexternal) <= 1e-10 )
+                        assert( transdist(testtorque_c2, torquenoexternal) <= 1e-10 )
+                        assert( transdist(testtorque_c3, torquenoexternal) <= 1e-10 )
                         
                         # compute the coriolis term: C_ij = 0.5*sum(dM_ij/dtheta_k + dM_ik/dtheta_j + dM_kj/dtheta_i)
                         def ComputeDynamicsM(body,dofvalues):
                             body.SetDOFValues(dofvalues,range(body.GetDOF()),checklimits=True)
-                            tb = body.ComputeInverseDynamics(zeros(body.GetDOF()))
                             M = zeros((body.GetDOF(),body.GetDOF()))
                             for index in range(body.GetDOF()):
                                 testaccel = zeros(body.GetDOF())
                                 testaccel[index] = 1.0
-                                M[:,index] = body.ComputeInverseDynamics(testaccel)-tb
+                                M[:,index] = body.ComputeInverseDynamics(testaccel)
                             return M
 
                         body.SetDOFVelocities(zeros(body.GetDOF()),[0,0,0],[0,0,0],checklimits=True)
@@ -486,7 +495,7 @@ class TestKinematics(EnvironmentSetup):
                         body.SetDOFValues(dofvaluesnew,range(body.GetDOF()),checklimits=True)
                         body.SetDOFVelocities(dofvelnew,*link0vel,checklimits=True)    
                         env.GetPhysicsEngine().SetGravity(2*gravity)
-                        torquegravity2 = body.ComputeInverseDynamics(zeros(body.GetDOF()))-torquenogravity
+                        torquegravity2 = body.ComputeInverseDynamics(None)-torquenogravity
                         assert(transdist(2*torquegravity,torquegravity2) <= g_epsilon)
 
                         # check that diff(M) - 2*C is skew symmetric
@@ -495,9 +504,9 @@ class TestKinematics(EnvironmentSetup):
                             Mdot += Mpartials[i]*dofvelnew[i]                            
                         SkewSymmetric = Mdot-2*C
                         assert(transdist(SkewSymmetric,-transpose(SkewSymmetric)) <= g_epsilon)
-
-                        # compute potential energy with respect to gravity
+                        
                         env.GetPhysicsEngine().SetGravity(gravity)
+                        # compute potential energy with respect to gravity
                         def ComputePotentialEnergy(body,dofvalues):
                             gravity = body.GetEnv().GetPhysicsEngine().GetGravity()
                             body.SetDOFValues(dofvalues,range(body.GetDOF()),checklimits=True)
@@ -514,7 +523,8 @@ class TestKinematics(EnvironmentSetup):
                             PEdiff = (ComputePotentialEnergy(body,dofvaluesnew+testdelta)-ComputePotentialEnergy(body,dofvaluesnew-testdelta))/(2*deltastep)
                             gravitypartials.append(PEdiff)
                         assert( transdist(-torquegravity, gravitypartials) < 0.1*deltastep*len(gravitypartials))
-                        
+                        assert( transdist(torquegravity, testtorque_e-testtorque_e2) <= 1e-10 )
+
     def test_initkinbody(self):
         self.log.info('tests initializing a kinematics body')
         with self.env: